# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardMetricTone = typeof DashboardMetricTone[keyof typeof DashboardMetricTone];


export const DashboardMetricTone = {
  positive: 'positive',
  neutral: 'neutral',
  attention: 'attention',
} as const;

export interface DashboardMetric {
  label: string;
  value: string;
  change: string;
  tone: DashboardMetricTone;
}

export type ActivityEventTone = typeof ActivityEventTone[keyof typeof ActivityEventTone];


export const ActivityEventTone = {
  positive: 'positive',
  neutral: 'neutral',
  attention: 'attention',
} as const;

export interface ActivityEvent {
  id: number;
  title: string;
  detail: string;
  time: string;
  tone: ActivityEventTone;
}

export interface DashboardSummary {
  role: string;
  headline: string;
  metrics: DashboardMetric[];
  activity: ActivityEvent[];
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  active: 'active',
  paused: 'paused',
  closed: 'closed',
} as const;

export interface Job {
  id: number;
  title: string;
  company: string;
  location: string;
  type: string;
  salary: string;
  posted: string;
  applicants: number;
  status: JobStatus;
}

export interface JobInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  location: string;
  type: string;
  salary: string;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  jobTitle: string;
  company: string;
  candidate: string;
  submitted: string;
  status: ApplicationStatus;
  match: number;
}

export interface ApplicationInput {
  jobId: number;
  candidate: string;
}

export type ApplicationStatusUpdateStatus = typeof ApplicationStatusUpdateStatus[keyof typeof ApplicationStatusUpdateStatus];


export const ApplicationStatusUpdateStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface ApplicationStatusUpdate {
  status: ApplicationStatusUpdateStatus;
}

export type PlatformUserRole = typeof PlatformUserRole[keyof typeof PlatformUserRole];


export const PlatformUserRole = {
  seeker: 'seeker',
  recruiter: 'recruiter',
  admin: 'admin',
} as const;

export type PlatformUserStatus = typeof PlatformUserStatus[keyof typeof PlatformUserStatus];


export const PlatformUserStatus = {
  pending: 'pending',
  accepted: 'accepted',
  removed: 'removed',
} as const;

export interface PlatformUser {
  id: number;
  name: string;
  email: string;
  role: PlatformUserRole;
  joined: string;
  status: PlatformUserStatus;
}

export type UserStatusUpdateStatus = typeof UserStatusUpdateStatus[keyof typeof UserStatusUpdateStatus];


export const UserStatusUpdateStatus = {
  pending: 'pending',
  accepted: 'accepted',
  removed: 'removed',
} as const;

export interface UserStatusUpdate {
  status: UserStatusUpdateStatus;
}

export type GetDashboardParams = {
role: GetDashboardRole;
};

export type GetDashboardRole = typeof GetDashboardRole[keyof typeof GetDashboardRole];


export const GetDashboardRole = {
  seeker: 'seeker',
  recruiter: 'recruiter',
  admin: 'admin',
} as const;

export type ListJobsParams = {
search?: string;
location?: string;
type?: string;
};

export type ListApplicationsParams = {
role: ListApplicationsRole;
status?: string;
};

export type ListApplicationsRole = typeof ListApplicationsRole[keyof typeof ListApplicationsRole];


export const ListApplicationsRole = {
  seeker: 'seeker',
  recruiter: 'recruiter',
  admin: 'admin',
} as const;

export type ListUsersParams = {
status?: string;
};