# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Role dashboard summaries
  - name: jobs
    description: Job listings
  - name: applications
    description: Seeker applications and recruiter pipeline
  - name: administration
    description: User moderation and platform controls
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary
      parameters:
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [seeker, recruiter, admin]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Job listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications for the current role
      parameters:
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [seeker, recruiter, admin]
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply to a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Accept, reject, or keep an application pending
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /users:
    get:
      operationId: listUsers
      tags: [administration]
      summary: List platform users
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Platform users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformUser"
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [administration]
      summary: Accept or remove a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated platform user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformUser"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DashboardSummary:
      type: object
      required: [role, headline, metrics, activity]
      properties:
        role:
          type: string
        headline:
          type: string
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/DashboardMetric"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEvent"
    DashboardMetric:
      type: object
      required: [label, value, change, tone]
      properties:
        label: { type: string }
        value: { type: string }
        change: { type: string }
        tone:
          type: string
          enum: [positive, neutral, attention]
    ActivityEvent:
      type: object
      required: [id, title, detail, time, tone]
      properties:
        id: { type: integer }
        title: { type: string }
        detail: { type: string }
        time: { type: string }
        tone:
          type: string
          enum: [positive, neutral, attention]
    Job:
      type: object
      required: [id, title, company, location, type, salary, posted, applicants, status]
      properties:
        id: { type: integer }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        type: { type: string }
        salary: { type: string }
        posted: { type: string }
        applicants: { type: integer }
        status:
          type: string
          enum: [active, paused, closed]
    JobInput:
      type: object
      required: [title, company, location, type, salary]
      properties:
        title: { type: string, minLength: 1 }
        company: { type: string, minLength: 1 }
        location: { type: string, minLength: 1 }
        type: { type: string }
        salary: { type: string }
    Application:
      type: object
      required: [id, jobTitle, company, candidate, submitted, status, match]
      properties:
        id: { type: integer }
        jobTitle: { type: string }
        company: { type: string }
        candidate: { type: string }
        submitted: { type: string }
        status:
          type: string
          enum: [pending, accepted, rejected]
        match: { type: integer }
    ApplicationInput:
      type: object
      required: [jobId, candidate]
      properties:
        jobId: { type: integer }
        candidate: { type: string }
    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, accepted, rejected]
    PlatformUser:
      type: object
      required: [id, name, email, role, joined, status]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [seeker, recruiter, admin]
        joined: { type: string }
        status:
          type: string
          enum: [pending, accepted, removed]
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, accepted, removed]