# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventTone } from './activityEventTone';

export interface ActivityEvent {
  id: number;
  title: string;
  detail: string;
  time: string;
  tone: ActivityEventTone;
}