# lib/api-zod/src/generated/types/applicationStatusUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatusUpdateStatus } from './applicationStatusUpdateStatus';

export interface ApplicationStatusUpdate {
  status: ApplicationStatusUpdateStatus;
}