# lib/api-zod/src/generated/types/dashboardMetric.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardMetricTone } from './dashboardMetricTone';

export interface DashboardMetric {
  label: string;
  value: string;
  change: string;
  tone: DashboardMetricTone;
}