# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEvent } from './activityEvent';
import type { DashboardMetric } from './dashboardMetric';

export interface DashboardSummary {
  role: string;
  headline: string;
  metrics: DashboardMetric[];
  activity: ActivityEvent[];
}