# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  location: string;
  type: string;
  salary: string;
}