# lib/api-zod/src/generated/types/platformUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformUserRole } from './platformUserRole';
import type { PlatformUserStatus } from './platformUserStatus';

export interface PlatformUser {
  id: number;
  name: string;
  email: string;
  role: PlatformUserRole;
  joined: string;
  status: PlatformUserStatus;
}