# lib/api-zod/src/generated/types/userStatusUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatusUpdateStatus } from './userStatusUpdateStatus';

export interface UserStatusUpdate {
  status: UserStatusUpdateStatus;
}